+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - `QrImage` added to handle image bits.
 - `QrCode.make()` removed
 - `QrCode.isDark()` moved to `QrImage`
+- Added alpha numeric mode support via `addAlphaNumeric`
 
 ## 2.1.0